Change order of SetDisplayDevice(), HWRestore(), UnbindGART() and RestoreBIOSMemSize() to be exactly opposite to the Save procedure in EnterVT() (Matthias Hopf, Alan Hourihane). Fixed server crash on reset when a structure allocated in PreInit() was freed on CloseScreen(). Fixed ring buffer lock ups that happened because the structure that contained ringbuffer data was not zeroed after allocation. Fixed numerous warnings due to signed unsigned comparisons.

+++ src/i810_memory.c
@@ -55,7 +55,7 @@ SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 int
 I810AllocLow(I810MemRange * result, I810MemRange * pool, int size)
 {
-   if (size > pool->Size)
+   if (size > (long)pool->Size)
       return 0;
 
    pool->Size -= size;
@@ -69,7 +69,7 @@ I810AllocLow(I810MemRange * result, I810MemRange * pool, int size)
 int
 I810AllocHigh(I810MemRange * result, I810MemRange * pool, int size)
 {
-   if (size > pool->Size)
+   if (size > (long)pool->Size)
       return 0;
 
    pool->Size -= size;
